接続部分を外部化． 接続を試みている最中かどうかをstaticでクラス内に持てるようになったので，接続中であるのにもう一度接続しようとするのを回避できるようになった
